requirements.txt
# MCP Server Core
mcp>=1.0.0

# AI Model Integration
anthropic>=0.39.0

# HTTP Client for Mattermost API
requests>=2.31.0

# Mattermost Integration  
mattermostdriver>=7.3.2

# Async Support
aiohttp>=3.9.1
websockets>=12.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Optional: Future model providers
# openai>=1.3.0
# google-generativeai>=0.3.0